naturebot/requirements.txt
google-genai
google-api-core